+++ sdk/src/org.graalvm.polyglot/src/org/graalvm/polyglot/Value.java
@@ -1009,52 +1009,52 @@ public final class Value extends AbstractValue {
     }
 
     /**
-     * Returns {@code true} if this value provides a {@linkplain #getStaticReceiver() static
-     * receiver}. A static receiver represents the static or class-level members associated with
-     * this value's type, such as static fields or methods.
+     * Returns {@code true} if this value is a {@linkplain #isMetaObject() meta object} that
+     * provides a {@linkplain #getStaticScope() static scope}. A static scope represents the static
+     * or class level members associated with the type described by this meta object, such as static
+     * fields or methods.
      *
      * @throws IllegalStateException if the context is already {@linkplain Context#close() closed}
      * @throws PolyglotException if a guest language error occurs during execution
-     * @see #getStaticReceiver()
+     * @see #isMetaObject()
+     * @see #getStaticScope()
      * @since 25.1
      */
-    public boolean hasStaticReceiver() {
-        return dispatch.hasStaticReceiver(this.context, receiver);
+    public boolean hasStaticScope() {
+        return dispatch.hasStaticScope(this.context, receiver);
     }
 
     /**
-     * Returns the static receiver associated with this value. A static receiver is an object that
-     * exposes static members, members whose values or behaviors are independent of any particular
-     * instance of this value. If this value is a {@link #isMetaObject() metaobject}, its static
-     * receiver is the same object that would be returned for an instance of the type represented by
-     * that metaobject, it exposes the static members of the represented type rather than static
-     * members of the metaobject itself.
+     * Returns the static scope associated with this value. This value must be a
+     * {@linkplain #isMetaObject() meta-object}. A static scope is an object that exposes static
+     * members, members whose values or behavior are independent of any particular instance.
      * <p>
-     * The returned static receiver can be used to access static members using
+     * The returned static scope can be used to access static members using
      * {@link #getMember(String)}, {@link #getMemberKeys()}, or
      * {@link #invokeMember(String, Object...)}.
      * <p>
-     * The returned static receiver is always expected to provide {@link #hasMembers() members},
-     * representing the receiver's static context.
+     * The returned static scope is always expected to provide {@link #hasMembers() members},
+     * representing the static context.
      * <p>
      * <b>Examples:</b>
+     * </p>
      * <ul>
-     * <li>For a Java class instance, the static receiver exposes the class's static fields and
-     * methods.</li>
-     * <li>For a Python object, the static receiver exposes class-level attributes and methods,
-     * effectively corresponding to the members provided by the Python metaobject.</li>
+     * <li>In Java, the static scope exposes static fields and methods of a class.</li>
+     * <li>In Python, the static scope exposes class-level attributes and methods, effectively
+     * corresponding to the members provided by the Python metaobject.</li>
      * </ul>
      *
      * @throws UnsupportedOperationException if and only if this value does not
-     *             {@linkplain #hasStaticReceiver() have a static receiver}
+     *             {@linkplain #hasStaticScope() have a static scope}
      * @throws IllegalStateException if the context is already {@linkplain Context#close() closed}
      * @throws PolyglotException if a guest language error occurs during execution
-     * @see #hasStaticReceiver()
-     * @see #getMemberKeys()
+     * @see #hasStaticScope()
+     * @see #isMetaObject()
+     * @see #hasMembers()
      * @since 25.1
      */
-    public Value getStaticReceiver() {
-        return (Value) dispatch.getStaticReceiver(this.context, receiver);
+    public Value getStaticScope() {
+        return (Value) dispatch.getStaticScope(this.context, receiver);
     }
 
     // executable